updated ae/graphics

+++ src/ae/ui/element.h
@@ -45,11 +45,11 @@ class _Element {
 		virtual void HandleInput(bool Pressed);
 		_Element *GetClickedElement();
 
-		void UpdateChildrenOffset(const glm::ivec2 &Update) { ChildrenOffset += Update; CalculateChildrenBounds(); }
+		void UpdateChildrenOffset(const glm::vec2 &Update) { ChildrenOffset += Update; CalculateChildrenBounds(); }
 		void CalculateChildrenBounds();
 
 		void SetDebug(int Debug);
-		void SetOffset(const glm::ivec2 &Offset) { this->Offset = Offset; CalculateBounds(); }
+		void SetOffset(const glm::vec2 &Offset) { this->Offset = Offset; CalculateBounds(); }
 		void SetWidth(int Width) { Size.x = Width; CalculateBounds(); }
 		void SetHeight(int Height) { Size.y = Height; CalculateBounds(); }
 
@@ -70,8 +70,8 @@ class _Element {
 		// Layout
 		_Bounds Bounds;
 		_Alignment Alignment;
-		glm::ivec2 Size;
-		glm::ivec2 Offset;
+		glm::vec2 Size;
+		glm::vec2 Offset;
 
 		// Input
 		_Element *HitElement;
@@ -80,7 +80,7 @@ class _Element {
 
 		// Children
 		std::vector<_Element *> Children;
-		glm::ivec2 ChildrenOffset;
+		glm::vec2 ChildrenOffset;
 
 	protected:
